\name{bigExplore-package}
\alias{bigExplore-package}
\alias{bigExplore}
\docType{package}
\title{
Large data set exploration
~~ bigExplore ~~
}
\description{
Large data set exploration, outlier detection. and linear regression diagnostics using big.matrix underlying data sets.
}
\details{
\tabular{ll}{
Package: \tab bigExplore\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-20\cr
License: \tab GPL\cr
}

}
\author{
Alex Ustian

Maintainer: Alex Ustian <alex.l.ustian@gmail.com>
}

